# docs/swagger.yaml
definitions:
  model.NewURL:
    properties:
      longUrl:
        type: string
    required:
    - longUrl
    type: object
  model.URL:
    properties:
      createdAt:
        description: TODO still as sting on swagger
        type: string
      externalId:
        type: string
      longUrl:
        type: string
    type: object
  urlErrors.GenericError:
    properties:
      statusCode:
        type: integer
    type: object
info:
  contact: {}
paths:
  /urls:
    post:
      consumes:
      - application/json
      description: Add a new URL
      operationId: add-url
      parameters:
      - description: URL to be added
        in: body
        name: url
        required: true
        schema:
          $ref: '#/definitions/model.NewURL'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.URL'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/urlErrors.GenericError'
      summary: Add a new URL
  /urls/:
    get:
      consumes:
      - application/json
      description: get all available short urls
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.URL'
            type: array
      summary: get all available short urls
      tags:
      - example
  /urls/{externalId}/visits:
    get:
      description: Count the number of visits for a URL based on its external ID
      operationId: count-url-visits
      parameters:
      - description: External ID of the URL
        in: path
        name: externalId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: visits
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/urlErrors.GenericError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/urlErrors.GenericError'
      summary: Count URL visits
swagger: "2.0"